feat: separa animação em um componente
